Move source code for tools to separate directory
... and move cmake file for QueryParser into parser directory
