+++ src/realm/exec/importer.cpp
@@ -25,7 +25,7 @@
 #include <vector>
 
 #include <realm/util/assert.hpp>
-#include <realm/importer.hpp>
+#include "importer.hpp"
 
 using namespace realm;
 
@@ -532,8 +532,9 @@ nextfield:
                 std::string s = payload[payload.size() - 1][0];
                 if (s.length() > 100)
                     s = s.substr(0, 100);
-                sprintf(buf, "Wrong number of delimitors around line %lld (+|- 3) in csv file. First few characters "
-                             "of line: %s",
+                sprintf(buf,
+                        "Wrong number of delimitors around line %lld (+|- 3) in csv file. First few characters "
+                        "of line: %s",
                         static_cast<unsigned long long>(m_row - 1), s.c_str());
                 throw std::runtime_error(buf);
             }
