+++ src/realm/exec/importer_tool.cpp
@@ -24,8 +24,8 @@
 #include <iostream>
 #include <realm.hpp>
 #include <realm/utilities.hpp>
-#include <realm/importer.hpp>
 #include <cstdarg>
+#include "importer.hpp"
 
 using namespace realm;
